[Major-changes] reconfigured the whole fole structure
