+++ LMS/LMS/settings.py
@@ -72,6 +72,10 @@ TEMPLATES = [
 WSGI_APPLICATION = 'LMS.wsgi.application'
 
 
+
+#AUTH_USER_MODEL = ''
+
+
 # Database
 # https://docs.djangoproject.com/en/5.0/ref/settings/#databases
