+++ LMS/api/users/admin.py
@@ -1,0 +1,16 @@
+from django.contrib import admin
+from .models import *
+
+
+
+@admin.register(User)
+class UserAdmin(admin.ModelAdmin):
+    list_display = ('ID','DateCreated','UserID','Name','Email','Password','Address','display_role','display_Rolename')
+    fields = ('Name','Email','Password','Address','Role_id')
+
+
+
+@admin.register(Role)
+class RoleAdmin(admin.ModelAdmin):
+    list_display = ('RoleID','RoleName','Description')
+
