+++ LMS/api/users/models.py
@@ -1,0 +1,5 @@
+from django.db import models
+from .__models__.user import User
+from .__models__.roles import Role 
+
+
